scrapers/requirements.txt
scrapy==2.11.0
feedparser==6.0.10
kafka-python==2.0.2
beautifulsoup4==4.12.2
python-dateutil==2.8.2
schedule==1.2.0
lxml==5.1.0
